Feat: Only show fund expenditure button if the expenditure is locked and not funded
