+++ src/components/common/Expenditures/ExpenditureDetailsPage/ExpenditureAdvanceButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ExpenditureAdvanceButton';
